Style slick carousel and begin read docs prompt
