+++ frontend/src/components/Carousel/Carousel.scss
@@ -1,0 +1,13 @@
+.custom-slick-carousel {
+  margin-bottom: -2em;
+  margin-top: -2em;
+}
+
+.slick-track {
+  padding-bottom: 3.5em;
+  padding-top: 3.5em;
+}
+
+.slick-arrow {
+  display: none !important;
+}
