+++ frontend/src/components/Carousel/CarouselCard.scss
@@ -1,0 +1,35 @@
+@import '../../styles/mixins';
+@import '../../styles/variables';
+
+.carousel-card {
+  @include card;
+
+  background-color: $white;
+  border-radius: 20px;
+  display: flex;
+  flex-direction: column;
+  height: 325px;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 325px;
+  padding: 1em;
+  position: relative;
+  transition: transform 0.5s;
+
+  &:hover {
+    cursor: pointer;
+    transform: scale(1.05);
+  }
+
+  h3 {
+    color: white;
+    font-size: 1em;
+    font-weight: bold;
+    margin-bottom: -0.5em;
+    margin-top: auto;
+    text-align: center;
+  }
+}
+
+.carousel-card-gradient {
+}
